refactor: 스크롤뷰(네일 그리드) 분리 및 네일 선택(컨테이너) 컴포넌트 추가 [SN-59]
